Standardize test resources in OpenSim/Common/
