+++ OpenSim/Common/tests/resources/dataWithEformat.trc
