+++ OpenSim/Common/tests/resources/dataWithNaNsOfDifferentCases.trc
