+++ OpenSim/Common/tests/resources/dataWithNaNsWithSpaces.trc
