+++ OpenSim/Common/tests/resources/exampleFormat.trc
