+++ OpenSim/Common/tests/resources/obj1Bad.xml
