+++ OpenSim/Common/tests/resources/obj1Defaults.xml
