+++ OpenSim/Common/tests/resources/sampleOutputs.sto
