+++ OpenSim/Common/tests/resources/sampleOutputsSpatialVec.sto
