+++ OpenSim/Common/tests/resources/sampleOutputsVec3.sto
