+++ OpenSim/Common/tests/resources/std_cop_walking2_grfs.sto
