+++ OpenSim/Common/tests/resources/std_cop_walking5_grfs.sto
