+++ OpenSim/Common/tests/resources/std_walking2_analog.sto
