+++ OpenSim/Common/tests/resources/std_walking2_markers.trc
