+++ OpenSim/Common/tests/resources/std_walking5_markers.trc
