+++ OpenSim/Common/tests/resources/test.sto
