+++ OpenSim/Common/tests/resources/testDiff.sto
