+++ OpenSim/Common/tests/testAssertion.cpp
