+++ OpenSim/Common/tests/testRootSolver.cpp
@@ -35,7 +35,7 @@
 #include <OpenSim/Common/PropertyStrArray.h>
 #include <OpenSim/Common/PropertySet.h>
 #include <OpenSim/Common/RootSolver.h>
-#include "ExampleVectorFunctionUncoupledNxN.h"
+#include <OpenSim/Common/VectorFunctionUncoupledNxN.h>
 
 #include <OpenSim/Auxiliary/auxiliaryTestFunctions.h>
 #include <catch2/catch_all.hpp>
@@ -46,6 +46,86 @@
 
 using namespace OpenSim;
 using namespace std;
+
+
+namespace {
+
+    /**
+    * An abstract class for representing a vector function.
+    *
+    * A vector function is a relation between some number of independent
+    * variables and some number of dependent values such that for any particular
+    * set of independent variables the correct number of dependent variables is
+    * returned. Values of the function and its derivatives are obtained by
+    * calling the evaluate() method. The curve may or may not be finite or
+    * differentiable; the evaluate method returns values between
+    * Math::MINUS_INFINITY and Math::PLUS_INFINITY, or it returns Math::NAN
+    * (not a number) if the curve is not defined.
+    * Currently, functions of up to 3 variables (x,y,z) are supported.
+    *
+    * @author Frank C. Anderson
+    */
+    class ExampleVectorFunctionUncoupledNxN : public VectorFunctionUncoupledNxN {
+        OpenSim_DECLARE_CONCRETE_OBJECT(ExampleVectorFunctionUncoupledNxN,
+                                        VectorFunctionUncoupledNxN);
+
+    public:
+        // CONSTRUCTION
+        ExampleVectorFunctionUncoupledNxN() : VectorFunctionUncoupledNxN(1) {}
+        ExampleVectorFunctionUncoupledNxN(int aN) :
+            VectorFunctionUncoupledNxN(aN) {}
+        ExampleVectorFunctionUncoupledNxN(
+            const ExampleVectorFunctionUncoupledNxN &aVectorFunction) :
+                VectorFunctionUncoupledNxN(aVectorFunction) {}
+        virtual ~ExampleVectorFunctionUncoupledNxN() {}
+
+        // OPERATORS
+        ExampleVectorFunctionUncoupledNxN&
+        operator=(const ExampleVectorFunctionUncoupledNxN& source) {
+            if (&source != this) {
+                Super::operator=(source);
+            }
+            return *this;
+        };
+
+        // EVALUATE
+        void calcValue(const double *aX,double *rY, int aSize) override {
+            int N = getNX();
+
+            // COMMON PART
+            int i;
+            double sum;
+            double scale = 0.01;
+            for(sum=0.0,i=0;i<N;i++) {
+                sum += (double)i;
+            }
+            sum *= scale;
+
+            // UNIQUE PART
+            // Uncoupled-- each aY depends only on its corresponding aX.
+            double root;
+            for(i=0;i<N;i++) {
+                root = scale * (double)i;
+                // sin test function
+                rY[i] = sum * sin(aX[i] - root);
+                // parabolic test function
+                //rY[i] = sum *aX[i]*aX[i]*aX[i] - sum*root*root*root;
+            }
+        }
+
+        void calcValue(const Array<double> &aX,Array<double> &rY) override {
+            calcValue(&aX[0],&rY[0], aX.getSize());
+        }
+
+        void calcDerivative(const Array<double>&, Array<double>&,
+                const Array<int>&) override {
+            log_cout("ExampleVectorFunctionUncoupledNxN.calcDerivative(): "
+                     "not implemented.");
+        }
+
+    };
+}
+
 
 TEST_CASE("APDM Data Reader")
 {
