+++ OpenSim/Common/tests/testSTOFileAdapter.cpp
