+++ OpenSim/Common/tests/testScopeExit.cpp
