+++ OpenSim/Common/tests/testSmoothSegmentedFunctionFactory.cpp
