+++ OpenSim/Common/tests/testTRCFileAdapter.cpp
