+++ OpenSim/Examples/DataAdapter/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable(exampleTRCFileAdapter exampleTRCFileAdapter.cpp)
 target_link_libraries(exampleTRCFileAdapter ${Simbody_LIBRARIES} osimCommon)
 set_target_properties(exampleTRCFileAdapter PROPERTIES FOLDER "Examples")
-file(COPY ${CMAKE_SOURCE_DIR}/OpenSim/Common/Test/dataWithNaNsOfDifferentCases.trc
+file(COPY ${CMAKE_SOURCE_DIR}/OpenSim/Common/tests/resources/dataWithNaNsOfDifferentCases.trc
      DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 
 if(WITH_EZC3D)
